[Auth] Moving Authenticator classes to a new Authenticator namespace for better organization.
