+++ src/Valkyrja/Application/Env.php
@@ -17,12 +17,12 @@ use Twig\Extension\ExtensionInterface as TwigExtensionInterface;
 use Valkyrja\Application\Constant\ComponentClass;
 use Valkyrja\Application\Contract\Application;
 use Valkyrja\Application\Support\Component;
+use Valkyrja\Auth\Authenticator\Contract\Authenticator;
+use Valkyrja\Auth\Authenticator\SessionAuthenticator;
 use Valkyrja\Auth\Constant\RouteName;
 use Valkyrja\Auth\Constant\SessionId;
-use Valkyrja\Auth\Contract\Authenticator;
 use Valkyrja\Auth\Entity\Contract\User;
 use Valkyrja\Auth\Entity\User as UserEntity;
-use Valkyrja\Auth\SessionAuthenticator;
 use Valkyrja\Auth\Store\Contract\Store;
 use Valkyrja\Auth\Store\OrmStore;
 use Valkyrja\Broadcast\Broadcaster\Contract\Broadcaster;
