+++ src/Valkyrja/Auth/Authenticator/EncryptedJwtAuthenticator.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Auth;
+namespace Valkyrja\Auth\Authenticator;
 
 use Override;
 use Valkyrja\Auth\Data\Contract\AuthenticatedUsers;
