+++ src/Valkyrja/Auth/Authenticator/JwtAuthenticator.php
@@ -11,9 +11,10 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Auth;
+namespace Valkyrja\Auth\Authenticator;
 
 use Valkyrja\Auth\Constant\HeaderValue;
+use Valkyrja\Auth\Data;
 use Valkyrja\Auth\Data\Contract\AuthenticatedUsers;
 use Valkyrja\Auth\Entity\Contract\User;
 use Valkyrja\Auth\Exception\InvalidAuthenticationException;
@@ -21,9 +22,12 @@ use Valkyrja\Auth\Hasher\Contract\PasswordHasher;
 use Valkyrja\Auth\Store\Contract\Store;
 use Valkyrja\Http\Message\Constant\HeaderName;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
+use Valkyrja\Jwt\Manager\Contract\Jwt;
+
+use function is_string;
 
 /**
- * Class TokenAuthenticator.
+ * Class JwtAuthenticator.
  *
  * @author Melech Mizrachi
  *
@@ -31,13 +35,14 @@ use Valkyrja\Http\Message\Request\Contract\ServerRequest;
  *
  * @extends AbstractAuthenticator<U>
  */
-class TokenAuthenticator extends AbstractAuthenticator
+class JwtAuthenticator extends AbstractAuthenticator
 {
     /**
      * @param Store<U>        $store  The store
      * @param class-string<U> $entity The user entity
      */
     public function __construct(
+        protected Jwt $jwt,
         protected ServerRequest $request,
         Store $store,
         PasswordHasher $hasher,
@@ -84,8 +89,15 @@ class TokenAuthenticator extends AbstractAuthenticator
      */
     protected function getAuthenticatedUsersFromToken(string $token): AuthenticatedUsers|null
     {
+        $jwtPayload = $this->jwt->decode($token);
+        $users      = $jwtPayload['users'] ?? null;
+
+        if (! is_string($users)) {
+            throw new InvalidAuthenticationException('Invalid token structure. Expecting users');
+        }
+
         $unserializedUsers = unserialize(
-            $token,
+            $users,
             ['allowed_classes' => true]
         );
