+++ src/Valkyrja/Auth/Provider/ServiceProvider.php
@@ -15,19 +15,19 @@ namespace Valkyrja\Auth\Provider;
 
 use Override;
 use Valkyrja\Application\Env;
-use Valkyrja\Auth\Contract\Authenticator;
-use Valkyrja\Auth\EncryptedJwtAuthenticator;
-use Valkyrja\Auth\EncryptedTokenAuthenticator;
+use Valkyrja\Auth\Authenticator\Contract\Authenticator;
+use Valkyrja\Auth\Authenticator\EncryptedJwtAuthenticator;
+use Valkyrja\Auth\Authenticator\EncryptedTokenAuthenticator;
+use Valkyrja\Auth\Authenticator\JwtAuthenticator;
+use Valkyrja\Auth\Authenticator\SessionAuthenticator;
+use Valkyrja\Auth\Authenticator\TokenAuthenticator;
 use Valkyrja\Auth\Entity\Contract\User;
 use Valkyrja\Auth\Hasher\Contract\PasswordHasher;
 use Valkyrja\Auth\Hasher\PhpPasswordHasher;
-use Valkyrja\Auth\JwtAuthenticator;
-use Valkyrja\Auth\SessionAuthenticator;
 use Valkyrja\Auth\Store\Contract\Store;
 use Valkyrja\Auth\Store\InMemoryStore;
 use Valkyrja\Auth\Store\NullStore;
 use Valkyrja\Auth\Store\OrmStore;
-use Valkyrja\Auth\TokenAuthenticator;
 use Valkyrja\Container\Contract\Container;
 use Valkyrja\Container\Support\Provider;
 use Valkyrja\Crypt\Manager\Contract\Crypt;
