+++ tests/Unit/Auth/Provider/ServiceProviderTest.php
@@ -14,19 +14,19 @@ declare(strict_types=1);
 namespace Valkyrja\Tests\Unit\Auth\Provider;
 
 use PHPUnit\Framework\MockObject\Exception;
-use Valkyrja\Auth\Contract\Authenticator as Contract;
-use Valkyrja\Auth\EncryptedJwtAuthenticator;
-use Valkyrja\Auth\EncryptedTokenAuthenticator;
+use Valkyrja\Auth\Authenticator\Contract\Authenticator as Contract;
+use Valkyrja\Auth\Authenticator\EncryptedJwtAuthenticator;
+use Valkyrja\Auth\Authenticator\EncryptedTokenAuthenticator;
+use Valkyrja\Auth\Authenticator\JwtAuthenticator;
+use Valkyrja\Auth\Authenticator\SessionAuthenticator;
+use Valkyrja\Auth\Authenticator\TokenAuthenticator;
 use Valkyrja\Auth\Hasher\Contract\PasswordHasher;
 use Valkyrja\Auth\Hasher\PhpPasswordHasher;
-use Valkyrja\Auth\JwtAuthenticator;
 use Valkyrja\Auth\Provider\ServiceProvider;
-use Valkyrja\Auth\SessionAuthenticator;
 use Valkyrja\Auth\Store\Contract\Store;
 use Valkyrja\Auth\Store\InMemoryStore;
 use Valkyrja\Auth\Store\NullStore;
 use Valkyrja\Auth\Store\OrmStore;
-use Valkyrja\Auth\TokenAuthenticator;
 use Valkyrja\Crypt\Manager\Contract\Crypt;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Jwt\Manager\Contract\Jwt;
